requirements.txt
web3==6.11.1
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
eth-account==0.10.0
eth-typing==4.0.0
eth-utils==2.3.1 